+++ sdl2-hyper-sonic-drivers/src/HyperSonicDrivers/hardware/opl/woody/WoodyOPL.hpp
@@ -26,7 +26,7 @@ namespace HyperSonicDrivers::hardware
                 void writeReg(const uint16_t r, const uint16_t v) noexcept override;
 
             protected:
-                void generateSamples(int16_t* buffer, int length) noexcept override;
+                void generateSamples(int16_t* buffer, size_t length) noexcept override;
 
             private:
                 std::unique_ptr<woody::OPL> _opl = nullptr;
